feat: move lucia demo to the top level
