+++ apps/user/src/routes/(authed)/+layout.server.ts
@@ -1,0 +1,9 @@
+import type { PageServerLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
+
+export const load: PageServerLoad = async (event) => {
+  if (!event.locals.user) {
+    return redirect(302, '/login');
+  }
+  return { user: event.locals.user };
+};
