+++ apps/user/src/routes/(authed)/mypage/+page.svelte
@@ -7,6 +7,6 @@
 
 <h1>Hi, {data.user.username}!</h1>
 <p>Your user ID is {data.user.id}.</p>
-<form method='post' action='?/logout' use:enhance>
+<form method='post' action='/logout' use:enhance>
   <button>Sign out</button>
 </form>
