+++ apps/user/src/routes/+page.svelte
@@ -1,6 +1,6 @@
 <h1>Welcome to SvelteKit</h1>
 <p>Visit <a href='https://svelte.dev/docs/kit'>svelte.dev/docs/kit</a> to read the documentation</p>
 <div>
-  <a href='/demo/lucia'>lucia</a>
+  <a href='/mypage'>lucia</a>
   <a href='/paraglide'>paraglide</a>
 </div>
