--- apps/user/src/routes/demo/+page.svelte
@@ -1,2 +1,0 @@
-<a href='/demo/lucia'>lucia</a>
-<a href='/demo/paraglide'>paraglide</a>
