+++ apps/user/src/routes/login/+page.server.ts
@@ -16,7 +16,7 @@ const hashParams = {
 
 export const load: PageServerLoad = async (event) => {
   if (event.locals.user) {
-    return redirect(302, '/demo/lucia');
+    return redirect(302, '/');
   }
   return {};
 };
@@ -53,7 +53,7 @@ export const actions: Actions = {
     const session = await auth.createSession(sessionToken, existingUser.id);
     auth.setSessionTokenCookie(event, sessionToken, session.expiresAt);
 
-    return redirect(302, '/demo/lucia');
+    return redirect(302, '/');
   },
   register: async (event) => {
     const formData = await event.request.formData();
@@ -84,7 +84,7 @@ export const actions: Actions = {
       console.error(e);
       return fail(500, { message: 'An error has occurred' });
     }
-    return redirect(302, '/demo/lucia');
+    return redirect(302, '/');
   },
 };
