+++ apps/user/src/routes/logout/+server.ts
@@ -1,0 +1,12 @@
+import * as auth from '$lib/server/auth';
+import { fail, redirect } from '@sveltejs/kit';
+
+export async function POST(event) {
+  if (!event.locals.session) {
+    return fail(401);
+  }
+  await auth.invalidateSession(event.locals.session.id);
+  auth.deleteSessionTokenCookie(event);
+
+  return redirect(302, '/login');
+}
